requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
coverage==5.5
cryptography==41.0.4
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.2
django-cache-url==3.4.4
django-js-asset==2.2.0
django-markdownx==4.0.2
emojidb==0.0.3
environs==9.5.0
eyed3==0.9.7
filetype==1.2.0
idna==3.4
Markdown==3.4.4
marshmallow==3.20.1
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.6.3
urllib3==2.0.5
whitenoise==6.6.0
